nexus-pm/services/auth-service/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
mysqlclient==2.2.4
redis==5.0.4
python-decouple==3.8
gunicorn==22.0.0
